# openapi.yml
openapi: 3.0.3
info:
  title: Archethos API
  version: 1.0.0
paths:
...
  /api/v1/media/:
    get:
      operationId: media_list
      parameters:
      - in: query
        name: media_type
        schema:
          type: string
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      - in: query
        name: source_type
        schema:
          type: string
      - in: query
        name: tags
        schema:
          type: string
      - in: query
        name: uploaded_by
        schema:
          type: integer
      tags:
      - media
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMediaAssetList'
          description: ''
    post:
      operationId: media_create
      tags:
      - media
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaAsset'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MediaAsset'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/MediaAsset'
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaAsset'
          description: ''
...
components:
  schemas:
...
    MediaAsset:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        file:
          type: string
          format: uri
          nullable: true
        thumbnail_file:
          type: string
          format: uri
          nullable: true
        source:
          type: string
          format: uri
        thumbnail_url:
          type: string
          format: uri
        media_type:
          type: string
          readOnly: true
        source_type:
          type: string
          readOnly: true
        external_url:
          type: string
          format: uri
          readOnly: true
        external_id:
          type: string
          readOnly: true
        file_name:
          type: string
          readOnly: true
        mime_type:
          type: string
          readOnly: true
        title:
          type: string
          maxLength: 255
        alt_text:
          type: string
          maxLength: 255
        caption:
          type: string
          maxLength: 500
        description:
          type: string
        tags:
          type: string
        media_location:
          type: string
          readOnly: true
          nullable: true
        youtube_url:
          type: string
          format: uri
          writeOnly: true
        video_url:
          type: string
          format: uri
          writeOnly: true
        uploaded_by:
          type: integer
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - external_id
      - external_url
      - file_name
      - id
      - media_location
      - media_type
      - mime_type
      - source_type
      - uploaded_by
...
    PaginatedMediaAssetList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/MediaAsset'
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token